% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addhealth.R
\docType{data}
\name{AddHealth}
\alias{AddHealth}
\title{AddHealth}
\source{
\href{https://www.cpc.unc.edu/projects/addhealth/documentation}{The National Longitudinal Study of Adolescent to Adult Health}
}
\usage{
AddHealth
}
\description{
Add Health study has collected data of interest to investigators from many disciplines in the social and behavioral sciences and from many theoretical traditions.
}
\keyword{datasets}
