% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cttraffic.R
\docType{data}
\name{CTTraffic}
\alias{CTTraffic}
\title{CtTraffic}
\source{
\href{https://data.ct.gov/Public-Safety/Traffic-Stops-Racial-Profiling-Prohibition-Project/nahi-zqrt}{CTData.gov}
}
\usage{
CtTraffic
}
\description{
Connecticut Racial Profiling Prohibition Project Data
}
\keyword{datasets}
