% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_search.r
\name{data_search}
\alias{data_search}
\title{Data search by keyword}
\usage{
data_search(search_str, datasets = DATASETS)
}
\arguments{
\item{search_str}{a string, case insensitive}

\item{datasets}{a vector of the selected datasets to search. Values must be
in of the following supported datasets: \code{FRED}, \code{WDI},
\code{AddHealth}, \code{MarsCrater}, \code{CtTraffic}. Defaults to
searching all supported datasets.}
}
\value{
A dataframe w/ columns - \code{dataset}, \code{col_id},
\code{description}, \code{notes}
}
\description{
Find columns in supported datasets that relate to keyword
}
\examples{
data_search("birth")
data_search("birth", "AddHealth")
data_search("birth", c("AddHealth", "MarsCrater"))
}
