% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_description}
\alias{get_description}
\title{Get variable description}
\usage{
get_description(search_result, col_id)
}
\arguments{
\item{search_result}{a dataframe result from \code{\link{data_search}}}

\item{col_id}{a string of \emph{col_id} found with \code{\link{data_search}}}
}
\value{
a string
}
\description{
Find the text description of variables from supported datasets
}
\examples{
data_search("birth", "AddHealth") \%>\%
  get_description("H4TR10")
}
