% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.r
\name{get_note}
\alias{get_note}
\title{Get variable note}
\usage{
get_note(search_result, col_id)
}
\arguments{
\item{search_result}{a dataframe result from \code{\link{data_search}}}

\item{col_id}{a string of \emph{col_id} found with \code{\link{data_search}}}
}
\value{
a dataframe
}
\description{
Find the metadata of variables from chosen dataset
}
\examples{
data_search("birth", "AddHealth") \%>\%
  get_note("H4TR10")
}
