% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marscrater.R
\name{MarsCrater}
\alias{MarsCrater}
\title{MarsCrater}
\source{
\href{https://ieee-dataport.org/open-access/mars-crater-segmentation-dataset}{IEEEDataPort}
}
\usage{
MarsCrater
}
\description{
Global database for Mars that contains 378,540 craters statistically complete for diameters D ≥ 1 km
}
\keyword{datasets}
