% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacD-package.R
\docType{package}
\name{qacData-package}
\alias{qacData}
\alias{qacData-package}
\title{qacData: Datasets for Data Science Instruction}
\description{
This package provides data sets that can be used to illustrate data science techniques.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/rkabacoff/qacData}
  \item Report bugs at \url{https://github.com/rkabacoff/qacData/issues}
}

}
\author{
\strong{Maintainer}: Robert Kabacoff \email{rkabacoff@wesleyan.edu}

}
\keyword{internal}
