% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wdi_indicators.R
\name{WDI}
\alias{WDI}
\title{WDI}
\source{
\href{https://datacatalog.worldbank.org/dataset/world-development-indicators}{World Development Indicators}
}
\usage{
WDI
}
\description{
collection of development indicators, compiled from officially-recognized international sources
}
\keyword{datasets}
