% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qacdata-package.R
\docType{package}
\name{qacdata-package}
\alias{qacdata}
\alias{qacdata-package}
\title{qacdata: Concise Searching And Downloading Of Relevant Data}
\description{
A data package for searching and downloading from FRED, WDI, 
  Addhealth, MarsCrater, CtTraffic
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/fredcorpuz06/qacdata}
  \item Report bugs at \url{https://github.com/fredcorpuz06/qacdata/issues}
}

}
\author{
\strong{Maintainer}: Frederick Corpuz \email{fcorpuz@wesleyan.edu}

Authors:
\itemize{
  \item Kevin Koech \email{kkoech@wesleyan.edu}
}

}
\keyword{internal}
